fix: update redocly.yaml validation
