+++ __tests__/lint/deprecated-apiDefinitions/snapshot.js
@@ -18,6 +18,8 @@ Warning was generated by the configuration spec rule.
 
 Property \`lint\` is not expected here.
 
+Did you mean: env ?
+
 1 | apiDefinitions:
 2 |   main: ./openapi.yaml
 3 | lint:
